bench_local

Function get_cases

pub(crate) fn get_cases() -> Vec<Case>

Edit/add new patterns here. The summary will display the pattern text as the “name”. The static_s values below are results from an alternate tool